Try to make wheels not sink into the ground
As I haven't implemented animation of the "landing gears" for this
toy, I shouldn't claim that the gears compress. That will lead to the
simulator calculating how much they are compressed, and placing the
aircraft correspondingly lower, expecting the gear compression
animation to take care of moving the 3D modelled wheels
correspondingly up.