Reduce stutter
We don't need to try to optimise when to modify the altitude. It
apparently just need to incosistencies and stutter. Just do it each
time we get to the code that recalculates the altitude. (Do continue
to ignore very small changes in throttle and mixture input, though.)

Also tweak the epsilon parameters to the
SimConnect_AddToDataDefinition() calls. (Or maybe we should just use
zero epsilon always?)