Try to use the FREEZE_*_TOGGLE event things
Not sure, but I think I see less stutter now? I think the point in
transmitting these events is to make the sim freeze its own updates of
aircraft position and attitude. Which is what we want as we want to
set that completely ourselves.

Also subscribe to relative wind information, but that is not used yet.