Add animations for the skids to compress when on ground
For now it is hard to get it to actually rest on the ground, though,
as the Wasm module is doing its work all the time and it doesn't let
it. Working on that. My plan is that when the ignition switch is off,
the Wasm module will not do anything, and the aircraft will fall rest
naturally on the ground. (And fall out of the sky if you turn off the
ignition mid-air.) But I need to hack the code a bit first, so that it
unfreezes the sim from affecting the aircraft position and attitude
then.