Tweak contact points
Possibly the sim doesn't allow skids to animate? Pretend they are
wheels, let's see if that works, as we simulate the parking brake
always being on anyway.

(But there are lots of issues when landing or taking off. It is hard
to manage the change of state between where our WASM module controls
everything and where we want to let the keep the aircraft standing
freeely on the possibly uneven ground with landing gears appropriately
compressed differently.