Work on trying to make the landing "gear" animate properly
Use different animation names for each leg. Make the legs long so that
their compression is very visible. (Compression movement 2 ft.) Claim
they are wheels anyway in case the simulator animates compression only
of wheels?

Still, looks very wrong in the simulator. For some reason, when in
what is a mostly flat runway, the gears very oddly compressed, and all
in the air;)

It is possibe that the MSFS2Blend exporter version 0.40 (from
https://www.fsdeveloper.com/forum/resources/blender2msfs-toolkit.256/
) is causing some issues for me. Would the one from FlyByWire
(https://github.com/flybywiresim/msfs2blender2msfs/tree/rewrite) be
better? Why can't these people simply work with the upstream Khronos
glTF exporter/importer folks and submit PRs to that? Why all the
duplicated effort and forking?

When working on the gear problems it is best to not have the Wasm
module at all.

Please note that the Wasm module source code causes very werd
behaviour when "landing" or "taking off" at the moment. I will work on
that when I get tired of trying to make the landing gears behave
sanely.