Try to make the external camera saner
For some reason it now is initailly located just behind the aircraft,
so that it fills the whole screen. Not very useful.

I don't understand why the InitialXyz for cameras isin metres when
everything else in the .cfg files is in feet, and why the coordinates
are in X, Y, Z order. Consistency is apparently over-rated. Or is the
name just wrong, and they actually are in the normal Z (longitudinal),
X (lateral), Y (vertical) order? And experimenting with that seems to
have little effect anyway.

Drop the "Quick View" cameras, getting a single external camera
working sanely seems hard enough...