Slight refactoring
Don't have desiredState as a global variable. Cleaner that way. Also
rename it to directControl as it doesn't contain the "desired state"
as such, but the directly controllable (mutable) simulation variables.
It is just in this trivial toy aircraft that we control those dirctly
with user input, ignoring physical realities.

(In a more realistic similar aircraft, i.e. a multicopter, the
"desired state" would be the heading, vertical speed, and horizontal
speed, and the way to achieve that would be by modulating the
simulated rotor rotational speeds, which would then cause a simulated
change in the tilt and accelerations of the aircraft, which would then
cause a change in those mutable simulation variables, taking into
account intertia and moments of inertia, etc. Or something like that.)