Now landing and take-off work clearly better
I tried to add some "hysteresis" so that when the module has
determined that you are landing, it will commit to that. Ditto for
taking off. But still there is room for improvements, if you start to
take-off but then quickly go downwards again, you will end up below
ground. Fully automated landing and take-off would be the best. Later.

Another thing that needs to be taken care of is to make the verbose
logging not show up normally, without having to use ifdefs. Ideally
when not displaying logging, no formatting should take place either,
the compiler should be given the opportunity to realise that it is not
needed. Hmm. Maybe ifdefs (or actually, a constexpr bool flag) is the
best anyway.