Experiment with the subdivision modifier
Best to use it for separate objects that need to be rounded all over,
instead of for a whole fuselage with lots of creased edges (that will
stay sharp after the modifier does its work). The latter would mean
generating unnecessarily a lot of extra coplanar polygons surrounded
by the creased edges.