Tweak the landing gears trying to understand it better
Now the movement (compression) of each landing gear is exactly 1 foot,
and the "static" compression, which presumably means the compression
with zero fuel and payload, is exactly half of that.

Still the gear behaves weirdly. Two are exactly on the ground, two
sink into the ground. By swapping the ID parameters around in the
model XML file I can switch which gears are on the ground and which in
the ground, but I can't get all found to be on the ground... Not sure
what I am missing.

Hmm, now when writing this it occurs to me, maybe it is the WASM code
that is artificially keeping the aircraft at zero pitch, even if as
the rear gears are closer to the center of gravity they will be more
compressed and the aircraft thus should be at a slight nose-up pitch
when on the ground? Will have to check. There is also the static_pitch
parameter, maybe I need to set that to a non-zero value?

See also
https://devsupport.flightsimulator.com/questions/3069/connection-between-wheel-contact-point-numbers-gea.html