Add a simple mesh object with a few animated vertices
It was awfully hard to get it to work.

Here is a summary of what I had to do. Quite possibly one can do it in
a simpler way.

0) You have a mesh object. Create an armature (typically inside it) to
handle the animation of some vertices of the mesh.

1) There apparently has to be at least two bones in the armature, one
that you typically don't animate, and one that you animate. This was a
key point that it took a long time for me to get.

2) First make the armature the parent of the mesh, "with envelope
weights". This automatically creates vertex groups for the mesh
corresponding to the bones in the armature.

3) Then you can clear the parent relationship from the mesh, keeping
transformations.

4) Now assign each vertex in the mesh to one of the vertex groups
automatically created in step 2. Each vertex needs to be in a group.
Vertices that you don't want to move should be in the group for that
bone.

5) Drop the action into a NLA track, as usual.