Make the bendy thing actually bend in the direction of travel
Firstly, it is body-relative longitudinal (Z) velocity that I want,
not world-relative. But VELOCITY BODY Z simvar apparently i always
positive, probably because normal aircraft can't fly backwards... So
use the same ELEVATOR POSITION simvar that I use to make the aircraft
move forward and backward to make the thing bend in different
direction depending on the way it is flying.